% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Time.R.R
\name{Time.R_calc}
\alias{Time.R_calc}
\title{Calculate Time of Concentration and Lag Time}
\usage{
Time.R_calc(data, plot_watershed = FALSE, plot_formulas = FALSE)
}
\arguments{
\item{data}{A data frame containing watershed morphometric information.
Required columns: watershed ID, area (km2), mean slope (\%), basin length (km),
max elevation (masl), min elevation (masl), average elevation (masl),
curve number, manning coefficient, paved (TRUE/FALSE).}

\item{plot_watershed}{A logical value. If TRUE, plots a comparative ggplot for each watershed.}

\item{plot_formulas}{A logical value. If TRUE, plots a faceted ggplot to compare each formula for all watersheds.}
}
\value{
A data frame with calculated time of concentration and lag times.
}
\description{
Estimates time of concentration and lag time for watersheds using various methods.
}
